Configurable checks and refactoring
This commit adds a new configuration section to enable/disable checks.
The ConfigHandler class has been re-encoded in UTF-8, and the checks for
the existence of the file removed, Bukkit already does these checks. The
"EnabledFor" section is loaded as soon as the plugin is loaded and
stores in memory a Set<Material>, which is much faster for an EnumSet
than going through the whole list every time (complexity: O(1)). From my
experiments, the InventoryCreativeEvent is fired every time the
InventoryClickEvent is fired, so it's redundant. Restrictions on armor
stands and item frames did not prevent any duplication. This is a
partial solution as no restrictions were made with the off hand.
Instead, I replicated the behavior of the survival mode. Item cloning
verification is disabled by default, as it doesn't work in the
player's inventory. The compiled files have been removed, as they
are useless with the source files. Also, the code was no longer
compatible with older versions (see getItemInMainHand in 1.8).
Saved hotbars can also be dropped, so I've added an extra check.
Calls to ItemStack#getData() are now avoided in 1.13, as they
cause legacy support to load, even though the api-version is defined.